Vulkan refactor 1 of 2 (VulkanTexture and VulkanConstants)
