+++ filament/backend/src/vulkan/PlatformVkAndroid.cpp
@@ -16,6 +16,7 @@
 
 #include "vulkan/PlatformVkAndroid.h"
 
+#include "VulkanConstants.h"
 #include "VulkanDriverFactory.h"
 
 #include <utils/Panic.h>
@@ -29,10 +30,6 @@ using namespace bluevk;
 namespace filament {
 
 using namespace backend;
-
-// All vkCreate* functions take an optional allocator. For now we select the default allocator by
-// passing in a null pointer, and we highlight the argument by using the VKALLOC constant.
-constexpr VkAllocationCallbacks* VKALLOC = nullptr;
 
 Driver* PlatformVkAndroid::createDriver(void* const sharedContext) noexcept {
     ASSERT_PRECONDITION(sharedContext == nullptr, "Vulkan does not support shared contexts.");
