+++ filament/backend/src/vulkan/PlatformVkCocoa.mm
@@ -16,6 +16,7 @@
 
 #include "vulkan/PlatformVkCocoa.h"
 
+#include "VulkanConstants.h"
 #include "VulkanDriverFactory.h"
 
 #include <Cocoa/Cocoa.h>
@@ -34,9 +35,6 @@ using namespace bluevk;
 namespace filament {
 
 using namespace backend;
-// All vkCreate* functions take an optional allocator. For now we select the default allocator by
-// passing in a null pointer, and we highlight the argument by using the VKALLOC constant.
-constexpr VkAllocationCallbacks* VKALLOC = nullptr;
 
 Driver* PlatformVkCocoa::createDriver(void* sharedContext) noexcept {
     ASSERT_PRECONDITION(sharedContext == nullptr, "Vulkan does not support shared contexts.");
