+++ filament/backend/src/vulkan/VulkanContext.h
@@ -19,6 +19,7 @@
 
 #include "VulkanBinder.h"
 #include "VulkanCommands.h"
+#include "VulkanConstants.h"
 #include "VulkanDisposer.h"
 
 #include <backend/DriverEnums.h>
@@ -39,14 +40,6 @@
 
 namespace filament {
 namespace backend {
-
-// All vkCreate* functions take an optional allocator. For now we select the default allocator by
-// passing in a null pointer, and we highlight the argument by using the VKALLOC constant.
-constexpr VkAllocationCallbacks* VKALLOC = nullptr;
-
-// At the time of this writing, our copy of MoltenVK supports Vulkan 1.0 only.
-constexpr static const int VK_REQUIRED_VERSION_MAJOR = 1;
-constexpr static const int VK_REQUIRED_VERSION_MINOR = 0;
 
 struct VulkanRenderTarget;
 struct VulkanSurfaceContext;
