+++ filament/backend/src/vulkan/VulkanStagePool.h
@@ -60,7 +60,6 @@ private:
 
     // Store the current "time" (really just a frame count) and LRU eviction parameters.
     uint64_t mCurrentFrame = 0;
-    static constexpr uint32_t TIME_BEFORE_EVICTION = 3;
 };
 
 } // namespace filament
